Add files via upload
Some code is commented out, I've tested muxing of video and audio in a separate script vide_util.py. I added a try except block in play_youtube_video which will handle the not internet connection error,  a function which should be called if directory access was denied abd a function for combining the video and audio in a single file.